+++ src/components/componentShapeChart/index.ts
@@ -1,0 +1,6 @@
+import { register } from '../../shared/functions';
+import ComponentShapeChart from './componentShapeChart';
+
+register('component-shape-chart', ComponentShapeChart);
+
+export default ComponentShapeChart;