+++ src/components/pageUser/index.ts
@@ -2,5 +2,6 @@ import PageUser from './pageUser';
 import { register } from '../../shared/functions';
 import '../componentLineChart/index';
 import '../componentBarLineChart/index';
+import '../componentShapeChart/index';
 
 register('page-user', PageUser);